refactor: post => notice로 변경
